change func to method
