+++ pkg/insight/insightstore/deploymentstore_test.go
@@ -274,7 +274,7 @@ func TestGetDailyDeployments(t *testing.T) {
 			for i, meta := range tc.storedMeta {
 				fs.EXPECT().Get(gomock.Any(), gomock.Any()).Return(protojson.Marshal(meta))
 
-				keys := findPathFromMeta(meta, tc.args.from, tc.args.to)
+				keys := meta.ListChunkNames(tc.args.from, tc.args.to)
 				assert.Equal(t, len(tc.storedChunks[i]), len(keys))
 				for _, v := range tc.storedChunks[i] {
 					fs.EXPECT().Get(gomock.Any(), gomock.Any()).Return(protojson.Marshal(v))
@@ -580,154 +580,6 @@ func TestPutDeployment(t *testing.T) {
 			err := store.putDeployments(context.Background(), tc.args.projectID, tc.args.dailyDeployment, *model.InsightDeploymentVersion_V0.Enum(), now)
 
 			assert.Equal(t, tc.expectedErr, err)
-		})
-	}
-}
-
-func TestExtractDailyDeploymentFromChunk(t *testing.T) {
-	t.Parallel()
-
-	const (
-		testDateUnix = 1647820800 // 2022/03/21:00:00:00 UTC
-		dayUnix      = int64(time.Hour * 24 / time.Second)
-	)
-
-	testcases := []struct {
-		name                     string
-		chunk                    *model.InsightDeploymentChunk
-		from, to                 int64
-		expectedDailyDeployments []*model.InsightDeployment
-	}{
-		{
-			name: "SingleDeployments-SingleDeployments",
-			chunk: &model.InsightDeploymentChunk{
-				From: testDateUnix, To: testDateUnix + 2,
-				Deployments: []*model.InsightDeployment{
-					{CompletedAt: testDateUnix + 1},
-					{CompletedAt: testDateUnix + 2},
-				},
-			},
-			from: testDateUnix, to: testDateUnix + 2,
-			expectedDailyDeployments: []*model.InsightDeployment{
-				{CompletedAt: testDateUnix + 1},
-				{CompletedAt: testDateUnix + 2},
-			},
-		},
-		{
-			name: "MultipleDeployments-SingleDeployments",
-			chunk: &model.InsightDeploymentChunk{
-				From: testDateUnix - dayUnix, To: testDateUnix + dayUnix + 2,
-				Deployments: []*model.InsightDeployment{
-					{CompletedAt: testDateUnix - dayUnix + 1},
-					{CompletedAt: testDateUnix - dayUnix + 2},
-					{CompletedAt: testDateUnix + 1},
-					{CompletedAt: testDateUnix + 2},
-					{CompletedAt: testDateUnix + dayUnix + 1},
-					{CompletedAt: testDateUnix + dayUnix + 2},
-				},
-			},
-			from: testDateUnix, to: testDateUnix + dayUnix,
-			expectedDailyDeployments: []*model.InsightDeployment{
-				{CompletedAt: testDateUnix + 1},
-				{CompletedAt: testDateUnix + 2},
-			},
-		},
-		{
-			name: "MultipleDeployments-MultipleDeployments",
-			chunk: &model.InsightDeploymentChunk{
-				From: testDateUnix - dayUnix, To: testDateUnix + dayUnix + 2,
-				Deployments: []*model.InsightDeployment{
-					{CompletedAt: testDateUnix - dayUnix + 1},
-					{CompletedAt: testDateUnix - dayUnix + 2},
-					{CompletedAt: testDateUnix + 1},
-					{CompletedAt: testDateUnix + 2},
-					{CompletedAt: testDateUnix + dayUnix + 1},
-					{CompletedAt: testDateUnix + dayUnix + 2},
-				},
-			},
-			from: testDateUnix - dayUnix, to: testDateUnix + 2*dayUnix,
-			expectedDailyDeployments: []*model.InsightDeployment{
-				{CompletedAt: testDateUnix - dayUnix + 1},
-				{CompletedAt: testDateUnix - dayUnix + 2},
-				{CompletedAt: testDateUnix + 1},
-				{CompletedAt: testDateUnix + 2},
-				{CompletedAt: testDateUnix + dayUnix + 1},
-				{CompletedAt: testDateUnix + dayUnix + 2},
-			},
-		},
-	}
-
-	for _, tc := range testcases {
-		t.Run(tc.name, func(t *testing.T) {
-			got := extractDeploymentsFromChunk(tc.chunk, tc.from, tc.to)
-			assert.True(t, len(tc.expectedDailyDeployments) == len(got))
-			for i := range tc.expectedDailyDeployments {
-				assert.True(t, proto.Equal(tc.expectedDailyDeployments[i], got[i]))
-			}
-		})
-	}
-}
-
-func TestOverlap(t *testing.T) {
-	t.Parallel()
-
-	const (
-		testDateUnix = 1647820800 // 2022/03/21:00:00:00 UTC
-		dayUnix      = int64(time.Hour * 24 / time.Second)
-	)
-
-	testcases := []struct {
-		name           string
-		lhsFrom, lhsTo int64
-		rhsFrom, rhsTo int64
-		expected       bool
-	}{
-		{
-			name:     "No overlap 1",
-			lhsFrom:  testDateUnix - dayUnix,
-			lhsTo:    testDateUnix,
-			rhsFrom:  testDateUnix,
-			rhsTo:    testDateUnix + dayUnix,
-			expected: false,
-		},
-		{
-			name:     "No overlap 2",
-			lhsFrom:  testDateUnix,
-			lhsTo:    testDateUnix + dayUnix,
-			rhsFrom:  testDateUnix - dayUnix,
-			rhsTo:    testDateUnix,
-			expected: false,
-		},
-		{
-			name:     "Overlap same day",
-			lhsFrom:  testDateUnix,
-			lhsTo:    testDateUnix + dayUnix,
-			rhsFrom:  testDateUnix,
-			rhsTo:    testDateUnix + dayUnix,
-			expected: true,
-		},
-		{
-			name:     "Overlap",
-			lhsFrom:  testDateUnix - dayUnix,
-			lhsTo:    testDateUnix + dayUnix,
-			rhsFrom:  testDateUnix,
-			rhsTo:    testDateUnix + dayUnix,
-			expected: true,
-		},
-		{
-			name:     "Overlap contain",
-			lhsFrom:  testDateUnix - dayUnix,
-			lhsTo:    testDateUnix + dayUnix,
-			rhsFrom:  testDateUnix,
-			rhsTo:    testDateUnix,
-			expected: true,
-		},
-	}
-
-	for _, tc := range testcases {
-		t.Run(tc.name, func(t *testing.T) {
-			got := overlap(tc.lhsFrom, tc.lhsTo, tc.rhsFrom, tc.rhsTo)
-			assert.Equal(t, tc.expected, got)
 		})
 	}
 }
